refactor(declaration): wire e2eClockSchema through the module barrel (#4065)
structural-auditor: import the schema from ~/modules/declaration (barrel)
instead of the internal path; re-export e2eClockSchema + E2eClockInput from
index.ts. E2eClockInput is public API for the future Playwright fixture.
